net: bus: add af_bus address and af_bus socket address definitions
An AF_BUS socket can have many addresses associated. This allows to send
multicast packets on different domains. An af_bus address is an unsigned
64-bit value that contains two fields: a 16-bit prefix and a 48-bit client
address.

Each bus has an associated path name that uniquely identifies the bus.
So, a socket address is composed of the bus path and the peer address.

Clients can send unicast packets to each other and multicast to different
prefixes but they can only connect(2) to a special socket that owns the
bus an is known as the bus master.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>